docs: transfered the payoutRatios constant to a separate module
