+++ roulette/models/completeModel.js
@@ -1,1 +1,4 @@
 import './menuModel.js';
+
+//export let strUp = 0;
+//export let completeBet = 25;
