+++ roulette/models/consts/min_max.js
@@ -1,0 +1,3 @@
+// min_max
+export const minBet = 1;
+export const maxBet = 100;
