+++ roulette/models/menuModel.js
@@ -1,41 +1,11 @@
-// roulette-calculator
-export const minBet = 1;
-export const maxBet = 100;
+import { minBet } from './consts/min_max.js';
+import { maxBet } from './consts/min_max.js';
+import { payoutRatios } from './consts/payoutRatios.js';
 
 // пересекается ли трек с полем:
 export const trackOverField = true;
 
-//export let strUp = 0;
-//export let completeBet = 25;
 
-// коэффициенты выплат
-export const payoutRatios = {
-    numb: {
-        payout: 35,
-        position: 1,
-        maxbet: (maxBet * 1)
-    },
-    split: {
-        payout: 17,
-        position: 2,
-        maxbet: (maxBet * 2)
-    },
-    corner: {
-        payout: 8,
-        position: 4,
-        maxbet: (maxBet * 4)
-    },
-    street: {
-        payout: 11,
-        position: 3,
-        maxbet: (maxBet * 3)
-    },
-    six_line: {
-        payout: 5,
-        position: 6,
-        maxbet: (maxBet * 6)
-    }
-};
 // номера рулетки:
 export const rouletteNumber = {
     0: {
