+++ roulette/views/menu/menuView.html
@@ -60,6 +60,8 @@
         </div>
     </div>
 
+    <script src="../../models/consts/min_max.js" type="module"></script>
+    <script src="../../models/consts/payoutRatios.js" type="module"></script>
     <script src="../../models/menuModel.js" type="module"></script>
 </body>
